활동이력 조회 도메인 변경

+++ src/main/java/com/parkyc/devog/github/service/dto/ParsedContribute.java
@@ -3,6 +3,6 @@ package com.parkyc.devog.github.service.dto;
 public record ParsedContribute(
         String repository,
         String owner,
-        Long commitCount
+        int commitCount
 ) {
 }
